\name{bde9915}
\docType{data}
\alias{bde9915}

\title{Data Set: Working Paper \sQuote{bde9915}}

\description{
Data set employed for illustration in the working paper referenced below.
}

\usage{
bde9915
}

\format{
A list containing the following monthly series:
\itemize{
\item gipi: Italian industrial production index (1981:1-1996:12);
\item euprin: European price index of industrial production (1981:1-1993:1) (*);
\item metipi: Spanish production index of metal products (1980:1-1997:7).
}
(*) 1993:12 is reported in the reference document.
}

\note{
The authors of the reference working paper probably use additional variables such as 
holidays to obtain the results reported in the document.
}

\references{
Kaiser, R., and Maravall, A. (1999). 
\emph{Seasonal Outliers in Time Series}.
Banco de España, Servicio de Estudios. Working paper number 9915.
}

\keyword{datasets}
