\name{hicp}
\docType{data}
\alias{hicp}

\title{Data Set: Harmonised Indices of Consumer Prices}

\description{
Harmonised indices of consumer prices in the Euro area.
}

\usage{
hicp
}

\format{
A list containing the following monthly indices (2005=100).
%
\tabular{llc}{
\bold{Key} \tab \bold{Description} \tab \bold{Start} \cr 
ICP.M.U2.N.000000.4.INX \tab Overall index \tab 1990:1 \cr 
ICP.M.U2.N.010000.4.INX \tab Food and non-alcoholic beverages \tab 1990:1 \cr 
ICP.M.U2.N.011000.4.INX \tab Food \tab 1995:1 \cr 
ICP.M.U2.N.011200.4.INX \tab Meat \tab 1995:1 \cr 
ICP.M.U2.N.011300.4.INX \tab Fish \tab 1995:1 \cr 
ICP.M.U2.N.011600.4.INX \tab Fruit \tab 1995:1 \cr 
ICP.M.U2.N.011700.4.INX \tab Vegetables \tab 1995:1 \cr 
ICP.M.U2.N.FOODPR.4.INX \tab Processed food incl. alcohol and tobacco \tab 1990:1 \cr 
ICP.M.U2.N.FOODUN.4.INX \tab Unprocessed food \tab 1990:1 \cr 
ICP.M.U2.N.IGXE00.4.INX \tab Industrial goods excluding energy \tab 1990:1 \cr 
ICP.M.U2.N.NRGY00.4.INX \tab Energy \tab 1990:1 \cr 
ICP.M.U2.N.SERV00.4.INX \tab Services \tab 1990:1
}
%
All the series end in December 2013.
}

\source{
European Central Bank. 
Statistical Data Warehouse. \url{http://sdw.ecb.europa.eu/}.
%http://sdw.ecb.europa.eu/browse.do?node=2120778
%http://sdw.ecb.europa.eu/browseSelection.do?DATASET=0&sfl2=4&REF_AREA=308&sfl3=4&ICP_ITEM=000000&ICP_ITEM=010000&ICP_ITEM=011000&ICP_ITEM=011200&ICP_ITEM=011300&ICP_ITEM=011600&ICP_ITEM=011700&ICP_ITEM=FOODPR&ICP_ITEM=FOODUN&ICP_ITEM=IGXE00&ICP_ITEM=NRGY00&ICP_ITEM=SERV00&sfl4=4&ICP_SUFFIX=INX&node=2120778
}

\keyword{datasets}
