\name{ipi}
\docType{data}
\alias{ipi}

\title{Data Set: Industrial Production Indices}

\description{
Industrial production indices in the manufacturing sector
of European Monetary Union countries.
}

\usage{
ipi
}

\format{
A list containing monthly indices (2010=100) for the following countries.
%
\tabular{lclc}{
\bold{Country} \tab \bold{Start} \tab \bold{Country} \tab \bold{Start} \cr  
Belgium \tab 2000:1 \tab Luxembourg \tab 2000:1 \cr 
Germany \tab 1999:1 \tab Malta \tab 2000:1 \cr 
Estonia \tab 1999:1 \tab Netherlands \tab 1999:1 \cr 
Greece \tab 2000:1 \tab Austria \tab 1999:1 \cr 
Spain \tab 1999:1 \tab Portugal \tab 1999:1 \cr
France \tab 1999:1 \tab Slovenia \tab 1999:1 \cr
Italy \tab 1999:1 \tab Slovakia \tab 2000:1 \cr 
Cyprus \tab 2000:1 \tab Finland \tab 1999:1 \cr
Latvia \tab 2000:1 \tab
}
%
All the series end in December 2013.
}

\note{
Data base key: \dQuote{sts_inpr_m},
monthly short-term business statistics. Production in industry.

Gross data for Ireland is not available in Eurosta's data base, only seasonally 
or working day adjusted data are available for this country.
}

\source{
EUROSTAT. Statistical Office of the European Communities. 
\url{http://ec.europa.eu/eurostat}.
%\url{http://epp.eurostat.ec.europa.eu/portal/page/portal/statistics/search_database}.
}

\keyword{datasets}
